+++ packages/design-system/src/lib/components/dropdown-button/DropdownButton.tsx
@@ -1,5 +1,5 @@
-import { DropdownButton as DropdownButtonBS } from 'react-bootstrap'
-import { ReactNode } from 'react'
+import { DropdownButton as DropdownButtonBS, Dropdown } from 'react-bootstrap'
+import { ReactNode, ComponentType, ForwardRefExoticComponent } from 'react'
 import styles from './DropdownButton.module.scss'
 import { IconName } from '../icon/IconName'
 import { ButtonGroup } from '../button-group/ButtonGroup'
@@ -7,9 +7,14 @@ import { Icon } from '../icon/Icon'
 
 type DropdownButtonVariant = 'primary' | 'secondary'
 
+interface CustomToggleProps {
+  onClick: (event: React.MouseEvent<HTMLElement>) => void
+  [key: string]: unknown
+}
+
 interface DropdownButtonProps {
   id: string
-  title: string
+  title?: string
   variant?: DropdownButtonVariant
   icon?: IconName | ReactNode
   withSpacing?: boolean
@@ -18,6 +23,10 @@ interface DropdownButtonProps {
   disabled?: boolean
   children: ReactNode
   ariaLabel?: string
+  customToggle?: ComponentType<CustomToggleProps> | ForwardRefExoticComponent<CustomToggleProps>
+  customToggleClassname?: string
+  customToggleMenuClassname?: string
+  align?: 'end' | 'start'
 }
 
 export function DropdownButton({
@@ -30,8 +39,22 @@ export function DropdownButton({
   onSelect,
   disabled,
   ariaLabel,
+  customToggle,
+  customToggleClassname,
+  customToggleMenuClassname,
+  align,
   children
 }: DropdownButtonProps) {
+  // If customToggle is provided, use Dropdown instead of DropdownButtonBS
+  if (customToggle) {
+    return (
+      <Dropdown id={id} onSelect={onSelect} className={customToggleClassname} align={align}>
+        <Dropdown.Toggle as={customToggle} />
+        <Dropdown.Menu className={customToggleMenuClassname}>{children}</Dropdown.Menu>
+      </Dropdown>
+    )
+  }
+
   return (
     <DropdownButtonBS
       className={withSpacing ? styles.spacing : ''}
@@ -39,13 +62,14 @@ export function DropdownButton({
       title={
         <>
           {typeof icon === 'string' ? <Icon name={icon} /> : icon}
-          {title}
+          {title && title}
         </>
       }
       aria-label={ariaLabel || title}
       variant={variant}
       as={asButtonGroup ? ButtonGroup : undefined}
       disabled={disabled}
+      align={align}
       onSelect={onSelect}>
       {children}
     </DropdownButtonBS>