+++ packages/design-system/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# Non Published Changes
+
+- **DropdownButton:**
+  - Add `customToggle` prop to allow custom toggle components.
+  - Add `customToggleClassname` and `customToggleMenuClassname` props to allow custom styling of the custom toggle dropdown wrapper and menu.
+  - Add `align` prop to control the alignment of the dropdown menu.
+
 # [2.0.2](https://github.com/IQSS/dataverse-frontend/compare/@iqss/dataverse-design-system@2.0.1...@iqss/dataverse-design-system@2.0.2) (2024-06-23)
 
 ### Bug Fixes